Automatically initialise .env
Use .env.example as .env. Otherwise Artisan is unable to update the APP_KEY value on first startup. .env values can be override with either environment variables or with Pterodactyl's database-based config, so it doesn't really matter that we just use the example.

This didn't used to be necessary because .env used to be a host volume. Since .env is now part of a named volume, we need to initialise it ourselves.