Added a DNS Proxy Server
Added to the project a container with a DNS Proxy Server. This server
allows accessing the containers from the host using a fictional domain
name instead of their IP addresses. This is particularly useful because
the IP addresses change each time the containers are run. Also, using
the proxy it's not necessary to map the container ports to host ports in
order to access the containers, since each container is accessible
throught its hostname. With this, the host ports are free to use with
other applications and the containers are not exposed to the Internet
throught the host ports, improving the security