Created volume for redis cache
A new volume with a random hash was being created by the redis cache
service each time the container was executed.
I added a volume to the
service with a fixed name. This prevents the  creation of a new volume
on each execution and also gives the volume a meaningful name that can
be easily associated to the container/service when inspecting the
docker environment